feat: create hello file with function to print a string
This commit creates a file hello.go which has a main package and function
and a Hello function.

The main function only prints the return of the Hello using fmt.Println method.
The Hello function receives a parameter and concatenates it with "Hello, ".

When run the program with go hello.go, the output should be "Hello, world".
There's a unit test in hello_test.go file backing the code in hello.go.